cleanup/nas/pyproject.toml
[project]
name = "clean-nas"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Couchbase Build Team", email = "build-team@couchbase.com" }
]
dependencies = [
    "PyYAML>=6"
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
clean-nas = "clean_nas:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
